code cleanup for tasks
